chore: Extracted InterfaceState
